Upgrade internal count to int64, drop net/http import.
int on 32 bits build would easily overflow, which would cause issues. int64
overflows can be looked at later.

Recreate interface http.Flusher to drop the import on net/http, which is quite
heavy.

+++ circular_test.go
@@ -37,7 +37,7 @@ func TestBufferInternalState(t *testing.T) {
 		writeOk(t, b, item.written)
 		bytesWritten += len(item.written)
 		ut.AssertEqualIndex(t, i, item.buf, string(b.buf))
-		ut.AssertEqualIndex(t, i, bytesWritten, b.bytesWritten)
+		ut.AssertEqualIndex(t, i, int64(bytesWritten), b.bytesWritten)
 	}
 }
